# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handler.authRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  handler.authResponse:
    properties:
      token:
        type: string
      user:
        additionalProperties: true
        type: object
    type: object
...
  handler.changeEmailRequest:
    properties:
      email:
        type: string
    type: object
  handler.changePasswordRequest:
    properties:
      confirm_password:
        type: string
      new_password:
        type: string
      old_password:
        type: string
    type: object
...
  handler.errorResponse:
    properties:
      error:
        example: 请求失败
        type: string
    type: object
  handler.forgotPasswordRequest:
    properties:
      email:
        type: string
    type: object
...
  handler.messageResponse:
    properties:
      message:
        example: 操作成功
        type: string
    type: object
...
  handler.resetPasswordRequest:
    properties:
      confirm_password:
        type: string
      new_password:
        type: string
      token:
        type: string
    type: object
...
  handler.userResponse:
    properties:
      user:
        additionalProperties: true
        type: object
    type: object
  handler.verifyEmailRequest:
    properties:
      token:
        type: string
    type: object
...
info:
  contact: {}
  description: Markdown 图床批量转换平台 API
  title: PicBed Switcher API
  version: "1.0"
paths:
  /api/auth/email:
    put:
      consumes:
      - application/json
      parameters:
      - description: 邮箱信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.changeEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.userResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: 修改当前用户邮箱
      tags:
      - auth
  /api/auth/email/verification:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.messageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: 重发邮箱验证邮件
      tags:
      - auth
  /api/auth/email/verify:
    post:
      consumes:
      - application/json
      parameters:
      - description: 验证令牌
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.verifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.messageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: 验证邮箱
      tags:
      - auth
  /api/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.authRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.authResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: 用户名或邮箱登录
      tags:
      - auth
  /api/auth/password:
    put:
      consumes:
      - application/json
      parameters:
      - description: 密码信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.changePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.messageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: 修改当前用户密码
      tags:
      - auth
  /api/auth/password/forgot:
    post:
      consumes:
      - application/json
      parameters:
      - description: 邮箱信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.forgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.messageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: 发送密码重置邮件
      tags:
      - auth
  /api/auth/password/reset:
    post:
      consumes:
      - application/json
      parameters:
      - description: 重置密码信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.resetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.messageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: 使用重置令牌设置新密码
      tags:
      - auth
  /api/auth/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.userResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: 获取当前用户信息
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 注册信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.authRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.authResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: 用户注册
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"